Backend/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
black==20.8b1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
defusedxml==0.7.0rc1
diff-match-patch==20200713
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.4
django-cors-headers==3.5.0
django-import-export==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
et-xmlfile==1.0.1
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
isort==5.6.4
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.5
pathspec==0.8.1
psycopg2==2.8.6
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
python3-openid==3.2.0
pytz==2020.4
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
virtualenv==20.2.1
whitenoise==5.2.0
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
